requirements.txt
Django == 1.9.7
django_compressor == 2.0
django-libsass == 0.7
django-jsonfield >= 0.9.15
Pillow == 2.8.2
pyyaml == 3.11
